backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
langgraph
pypdf
pdfplumber
pandas
faiss-cpu
sentence-transformers
python-dotenv
unstructured
google-generativeai
pytest
